Add parameter sweep infrastructure and bilinear expansion factor sweep
- Create sweep config system with yaml format specifying parameter and values
- Add run_sweep.py to execute parameter sweeps sequentially
- Remove inline plotting from transcoding.py, save only training data
- Add plot_sweep.py to load and plot multiple experiments together
- Fix int() conversion for hidden_multiplier to support fractional values
- Fix directory paths to use project root for proper subdirectory execution
- Add bilinear_sweep.yaml config for hidden_multiplier sweep (0.25, 0.5, 1, 2, 4)
- Results show clear scaling: 4x multiplier achieves 28.7% variance explained vs 8.9% for 0.25x

+++ experiments/layer_prediction.py
@@ -126,7 +126,7 @@ bias = config_dict['layer_prediction']['bias']
 
 transcoder_cfg = Config(
     n_inputs=n_inputs,
-    n_hidden=n_inputs * hidden_multiplier,
+    n_hidden=int(n_inputs * hidden_multiplier),
     n_outputs=n_outputs,
     lr=learning_rate,
     device=device,
